mod_for_factory: adjust patch to boot-complete
mod_for_factory patches boot-complete since it disables the ui job,
meaning that login-prompt-visible will never happen. The correct
place to move "boot-complete" to is "started boot-services" - this
means all boot services are done.

system-services depends on boot-complete, so this means that without
ui, system-services will be run on started boot-services which is
what we want.

BUG=chromium-os:16079
TEST=none

Change-Id: I93b86eefe8df9da2e9fbff75aa1e9b80a1f9ad1b
Reviewed-on: http://gerrit.chromium.org/gerrit/2323
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Nick Sanders <nsanders@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>