Allow for new toolchains to be rolled out when sdk builder completes.
New toolchains should only be rolled out to developers running with
binary packages when the sdk-builder has completed and uploaded the
prebuilts. This ensures that we've vetted the toolchain properly
before it rolls out and also speeds up binary builds for developers.

This also provides consistency with the behavior of setup_board.

BUG=chromium-os:16363
TEST=Verify these 3 packages are only updated in --usepkg builds
     when the prebuilts are also updated.

Change-Id: I6be7c6c98c12130751d3d4440bec7c71afb1d4c2
Reviewed-on: http://gerrit.chromium.org/gerrit/2391
Reviewed-by: Raymes Khoury <raymes@chromium.org>
Tested-by: David James <davidjames@chromium.org>