CHROMIUM: build tools: Allow parallel use of cros_image_to_target.py
This change addresses a defect which prevents simultaneous use of
'cros_image_to_target.py'.

The issue is that the default invocation will always use the same port
number of HTTP communication.  This change, with default parameters,
will now start at the same port number, but will now instead iterate
through a number of ports (the number of ports is set to 10).

If no available port can be found, the script will fail with an
appropriate message.

If the user specifies a specific port, then only that port number will
be used; if the port is not available, the script will fail with an
approrpriate message.

BUG=chromium-os:17558
TEST=Uploaded two images simultaneously to different boards.

Change-Id: Icbbea12468fd766897daeecb0b7f1d3929f52a1c
Signed-off-by: Taylor Hutt <thutt@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/4110
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Terry Lambert <tlambert@chromium.org>