common.sh: add quotes where appropriate
Much of this file quotes paths which could possibly contain breaking
chars (e.g. spaces), but a few are missed.  In general, this probably
doesn't break as no one is adventurous enough to use spaces, but might
as well do it for the fun of it.  And to be consistent.  One of those.

BUG=None
TEST=ran make_chroot/enter_chroot/build_packages; seemed to be OK

Change-Id: Ic64d60790f50bb812c87a4672e2a24fafbccd749
Reviewed-on: http://gerrit.chromium.org/gerrit/5759
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>