mod_image_for_recovery.sh: don't silence stderr in non-verbose modes
The current code will silence stderr when operating in non-verbose mode
because it always turns on shell tracing with `set -x`.  Unfortunately,
this also ends up killing error messages from a bunch of places in the
script as well:
 - look for all the "1>&2" uses in mod_image_for_recovery.sh
 - look at missing error() override
 - look at duplication of die() just to undo things

A really simple example:
$ ./mod_image_for_recovery.sh
$
Did it work?  Or did I do something wrong?  Who knows!

So undo the stderr silencing and simply turn on `set -x` only when
requested (verbose mode).

BUG=None
TEST=forced some errors in non-verbose mode and saw no output before change, but saw it after change; made a working recovery image

Change-Id: I31578fba091e390a56a437af97782a621e2137fb
Reviewed-on: http://gerrit.chromium.org/gerrit/6904
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>