check_deps: handle "include" directives in ld.so.conf
The ld.so.conf file supports an "include" directive where it'll include
all the files matched by the following glob.  Since Gentoo has started
using this, we need to support parsing of it too.  So move the parsing
code into a dedicated function and recursively call ourselves when we
hit an include directive.

BUG=chromium-os:20636
TEST=`./check_deps /build/amd64-generic /bin/bash` now finds libs
TEST=`./build_image --board=x86-alex` still works

Change-Id: I8894ca42358d91d8f2ee6e95b47faf9334ccdd26
Reviewed-on: http://gerrit.chromium.org/gerrit/8494
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>