[crosutils] Make auth-testing cert CN be *.google.com
It seems that the intermittent failure we see on the login screen
is due to a sort of mixed-content error, where there's a failure
within the page to load some embedded resource due to a cert error.

Make the SSL cert we inject work for any google.com server to
work around this.

BUG=chromium-os:20323
TEST=login_CryptohomeMounted 50x, suite_Smoke

Change-Id: Ic8e5b9bec799cd19cccfeddd2990fe3a494d2184
Reviewed-on: http://gerrit.chromium.org/gerrit/8818
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>