image_to_usb.sh: autodetect a removable target device when none provided.
If no target device is provided, image_to_usb.sh will let the user
select one out of a list of autodetected devices. If only one device is
detected, it will be automatically selected.  Also improves the
descriptors shown for candidate/chosen target device(s), and prints
a noticeable warning when the target device does not appear to be
a USB/MMC one.

Also changed all fail/warning messages to use 'die' and 'warn',
respectively, for compliance with other scripts.  Slightly massaged
error/warning strings to be more compact and to-the-point.

BUG=chromium-os:25878
TEST=Executed in chroot environment with various combinations of flags
and connected devices

Change-Id: If248993b8e6f3bc8654c2c8f25f1e54e7899330d
Reviewed-on: https://gerrit.chromium.org/gerrit/15270
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Gilad Arnold <garnold@chromium.org>
Commit-Ready: Gilad Arnold <garnold@chromium.org>