common.sh: move output helper funcs up top the top
The shflags sourcing logic tries to use `die` before we've defined it, so
if there's an issue, we end up with:
	bash: die: command not found
and the shflags code doesn't actually get loaded.

So relocate these small helpers to the top so we don't have to worry about
when it's safe to use these things.

BUG=None
TEST=`rm lib/shflags/shflags && (. ./common.sh)` now exits properly

Change-Id: Ibdc268e6c081aa07679dc9fce76e5603b7217b20
Reviewed-on: https://gerrit.chromium.org/gerrit/15442
Reviewed-by: Brian Harring <ferringb@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>